# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Modzilla Private API - OpenAPI 3.0",
    "description": "This is a private API for the Modzilla project. It is not intended for public use yet.\n\nThis API is still in development and may change at any time.\n\nIf you have any questions or concerns, please contact us at simonvillafanee@gmail.com",
    "termsOfService": "https://swagger.io/terms/",
    "contact": {
      "email": "simonvillafanee@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/itssimmons/api.modzilla/blob/main/LICENSE"
    },
    "version": "0.1.0-alpha"
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "https://swagger.io"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/v1"
    },
    {
      "url": "http://127.0.0.1:5000/v1"
    }
  ],
  "tags": [
    {
      "name": "Rooms",
      "description": "Organize your chats using rooms"
    },
    {
      "name": "Chats",
      "description": "Chat with your friends"
    },
    {
      "name": "Users",
      "description": "Manage your account"
    }
  ],
  "paths": {
    "/channels/{room_id}/": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Get one room",
        "description": "Get one room, (Note: Getting one room is cached for an hour)",
        "operationId": "read_room",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Room ID",
              "type": "string"
            },
            "name": "room_id",
            "in": "path",
            "example": "53c38a2c-9640-4957-92d7-0d4400b2b9ac"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },

    "/chats/channel/{room_id}/": {
      "get": {
        "tags": ["Chats"],
        "summary": "Read all chats",
        "description": "Read all chats",
        "operationId": "read_chats",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Room ID",
              "type": "string"
            },
            "name": "room_id",
            "in": "path",
            "example": "53c38a2c-9640-4957-92d7-0d4400b2b9ac"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chat"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "tags": ["Chats"],
        "summary": "Send a message",
        "description": "Send a message",
        "operationId": "send_message",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Room ID",
              "type": "string"
            },
            "name": "room_id",
            "in": "path",
            "example": "53c38a2c-9640-4957-92d7-0d4400b2b9ac"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "The message content",
                    "example": "Hello? is someone there? 🔍"
                  },
                  "sender_id": {
                    "type": "number",
                    "description": "The unique identifier for the sender",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            }
          }
        }
      }
    },
    "/chats/channel/{room_id}/{chat_id}/react": {
      "post": {
        "tags": ["Chats"],
        "summary": "React to a message",
        "description": "React to a message",
        "operationId": "react_to_message",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Room ID",
              "type": "string"
            },
            "name": "room_id",
            "in": "path",
            "example": "53c38a2c-9640-4957-92d7-0d4400b2b9ac"
          },
          {
            "required": true,
            "schema": {
              "title": "Chat ID",
              "type": "string"
            },
            "name": "chat_id",
            "in": "path",
            "example": "35f6e2c9-8d20-4316-b9fc-77f8338343e6"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sender_id": {
                    "type": "number",
                    "description": "The unique identifier for the sender",
                    "example": 1
                  },
                  "emoji": {
                    "type": "string",
                    "description": "The emoji to react with",
                    "example": "👍"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            }
          }
        }
      }
    },

    "/players/": {
      "get": {
        "tags": ["Users"],
        "summary": "Read all users",
        "description": "Read all users",
        "operationId": "read_users",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/players/register/{username}": {
      "get": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "description": "Register a new user",
        "operationId": "register_player",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            },
            "name": "username",
            "in": "path",
            "example": "simon"
          },
          {
            "required": true,
            "schema": {
              "title": "Avatar",
              "type": "string"
            },
            "name": "avatar",
            "in": "query",
            "example": "https://ui-avatars.com/api/?background=005f5f&name=Simon&length=1"
          },
          {
            "required": true,
            "schema": {
              "title": "Color",
              "type": "string"
            },
            "name": "color",
            "in": "query",
            "example": "#005f5f"
          },
          {
            "required": true,
            "schema": {
              "title": "Room",
              "type": "string"
            },
            "name": "room",
            "in": "query",
            "example": "53c38a2c-9640-4957-92d7-0d4400b2b9ac"
          },
          {
            "required": true,
            "schema": {
              "title": "Status",
              "type": "string",
              "enum": ["online", "offline", "typing", "idle"]
            },
            "name": "status",
            "in": "query",
            "example": "online"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/players/{user_id}/": {
      "patch": {
        "tags": ["Users"],
        "summary": "Update a user",
        "description": "Update a user",
        "operationId": "update_player_status",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "User ID",
              "type": "number"
            },
            "name": "user_id",
            "in": "path",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "The status of the user",
                    "example": "online",
                    "enum": ["online", "offline", "typing", "idle"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBadRequest"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier for the user"
          },
          "username": {
            "type": "string",
            "description": "The username of the user"
          },
          "role": {
            "type": "string",
            "description": "The role of the user"
          },
          "avatar": {
            "type": "string",
            "description": "The URL of the user's avatar"
          },
          "color": {
            "type": "string",
            "description": "The color of the user's avatar"
          },
          "status": {
            "type": "string",
            "description": "The status of the user"
          }
        }
      },

      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the room"
          },
          "name": {
            "type": "string",
            "description": "The name of the room"
          },
          "description": {
            "type": "string",
            "description": "A description of the room"
          },
          "logo": {
            "type": "string",
            "description": "The URL of the room's logo"
          },
          "created_at": {
            "type": "string",
            "description": "The date and time the room was created"
          },
          "modified_at": {
            "type": "string",
            "description": "The date and time the room was last modified"
          }
        }
      },

      "Chat": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the chat"
          },
          "sender_id": {
            "type": "number",
            "description": "The unique identifier for the sender"
          },
          "message": {
            "type": "string",
            "description": "The message content"
          },
          "room_id": {
            "type": "string",
            "description": "The unique identifier for the room"
          },
          "created_at": {
            "type": "string",
            "description": "The date and time the chat was created"
          },
          "modified_at": {
            "type": "string",
            "description": "The date and time the chat was last modified"
          },
          "modified_id": {
            "type": "number",
            "description": "The unique identifier for the user who last modified the chat"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          }
        }
      },
      "Reaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the reaction"
          },
          "chat_id": {
            "type": "string",
            "description": "The unique identifier for the chat"
          },
          "sender_id": {
            "type": "number",
            "description": "The unique identifier for the sender"
          },
          "emoji": {
            "type": "string",
            "description": "The emoji used for the reaction"
          },
          "count": {
            "type": "number",
            "description": "The number of times the reaction was used"
          },
          "created_at": {
            "type": "string",
            "description": "The date and time the reaction was created"
          }
        }
      },

      "ApiOk": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "A message describing the result of the request",
            "example": "This pretends to be a OK message"
          }
        }
      },
      "ApiBadRequest": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request was successful",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "A message describing the result of the request",
            "example": "This pretends to be a BAD REQUEST message"
          }
        }
      }
    }
  }
}